SCT-1091 - Refactor component/page folder
